Refactor #90: extract pure tree mutations into tree-mutations.ts
Move pure tree-mutation logic out of the React hook useTreeOperations.ts
into a new src/utils/tree-mutations.ts (no React imports), leaving the hook
as thin commit/history orchestration (1417 -> 487 lines). Extracts
mergeNodesInDoc as the pure core of the old mergeNodes callback.

Module boundary: tree-utils holds position-addressed structural primitives;
tree-mutations holds id-addressed, intent-level editor operations composed
from them. Dependency is one-way (tree-mutations imports tree-utils).

Adds tree-mutations.test.ts with 47 isolated unit tests. The existing 189
useTreeOperations tests pass unchanged as the regression net.
